refactor: move Hazards view methods to view.py
