--- src/ramstk/views/gtk3/hazard_analysis/workview.pyi
@@ -1,50 +1,0 @@
-# Standard Library Imports
-from typing import Any, Dict, List, Tuple
-
-# Third Party Imports
-from treelib import Tree as Tree
-
-# RAMSTK Package Imports
-from ramstk.configuration import RAMSTKUserConfiguration as RAMSTKUserConfiguration
-from ramstk.logger import RAMSTKLogManager as RAMSTKLogManager
-from ramstk.views.gtk3 import Gtk as Gtk
-from ramstk.views.gtk3 import _ as _
-from ramstk.views.gtk3.widgets import RAMSTKPanel as RAMSTKPanel
-from ramstk.views.gtk3.widgets import RAMSTKWorkView as RAMSTKWorkView
-
-class HazOpsPanel(RAMSTKPanel):
-    _module: str = ...
-    _dic_attribute_keys: Any = ...
-    _dic_attribute_updater: Any = ...
-    _dic_row_loader: Any = ...
-    _title: Any = ...
-    def __init__(self) -> None: ...
-    def do_load_severity(
-        self, criticalities: Dict[int, Tuple[str, str, int]]
-    ) -> None: ...
-    def do_load_hazards(self, hazards: Dict[Any, Any]) -> None: ...
-    def do_load_probability(self, probabilities: List[str]) -> None: ...
-    def do_set_callbacks(self) -> None: ...
-    def _do_clear_panel(self) -> None: ...
-    def _on_insert(self, tree: Tree) -> None: ...
-    _parent_id: Any = ...
-    _record_id: Any = ...
-    def _on_row_change(self, selection: Gtk.TreeSelection) -> None: ...
-    def __do_set_properties(self) -> None: ...
-
-class HazOps(RAMSTKWorkView):
-    _module: str = ...
-    _tablabel: str = ...
-    _tabtooltip: str = ...
-    _lst_mnu_labels: Any = ...
-    _lst_tooltips: Any = ...
-    _pnlPanel: Any = ...
-    def __init__(
-        self, configuration: RAMSTKUserConfiguration, logger: RAMSTKLogManager
-    ) -> None: ...
-    def _do_request_calculate(self, __button: Gtk.ToolButton) -> None: ...
-    _record_id: Any = ...
-    def _do_set_record_id(self, attributes: Dict[str, Any]) -> None: ...
-    _parent_id: Any = ...
-    def _on_select_function(self, attributes: Dict[str, Any]) -> None: ...
-    def __make_ui(self) -> None: ...
